More generic changes
made it so that some of the files now automatically find the correct
winhue dir. also introduced a more universal add_ini file with a
customizable ini path depending on where the ini is in relationship to
the directory structure. files are now linked in the directory struture.
very cool!