Lots of changes!
Built a whole group scene builder.

changed the house scene builder so that it plays back scenes much faster

edited the green clouds file to now delete the stop task on startup,
just in case it was left behind from the last person leaving the house
when no animation was run. it was killing my animation when i came back
home! so frustrating!